Conflicts:
	app/controllers/omniauth_callbacks_controller.rb
	app/models/account/constants.rb

+++ app/models/user/callbacks.rb
@@ -22,7 +22,8 @@ class User < ActiveRecord::Base
   after_commit :clear_agent_caches, on: :create, :if => :agent?
   after_commit :update_agent_caches, on: :update
   after_commit :set_user_company, on: :create, :if => :customer_id?
-  after_commit :update_tickets_company_id, :update_user_company, on: :update, :if => :company_id_updated?
+  after_commit :update_user_company, on: :update
+  after_commit :update_tickets_company_id, on: :update, :if => :company_id_updated?
 
   after_commit :subscribe_event_create, on: :create, :if => :allow_api_webhook?
 
@@ -80,15 +81,20 @@ class User < ActiveRecord::Base
   end
 
   def set_user_company
-    user_companies.create(:company_id => customer_id)
+    user_companies.create(:company_id => customer_id, 
+                          :default => true, 
+                          :client_manager => privilege?(:client_manager))
   end
 
   def update_user_company
-    if customer_id.present?
-      user_companies.any? ? user_companies.first.update_attributes(:company_id => customer_id) : 
-                            set_user_company
-    elsif @model_changes.key?("customer_id")
-      user_companies.where(:company_id => @model_changes["customer_id"][0]).destroy_all
+    if company_id_updated? || privileges_updated?
+      if customer_id.present?
+        user_companies.any? ? user_companies.first.update_attributes(:company_id => customer_id,
+                                :default => true, :client_manager => privilege?(:client_manager)) : 
+                              set_user_company
+      elsif @model_changes.key?("customer_id")
+        user_companies.where(:company_id => @model_changes["customer_id"][0]).destroy_all
+      end
     end
   end
